Move btc setup classes to setup package
