+++ core/src/main/java/bisq/core/app/BisqExecutable.java
@@ -20,9 +20,9 @@ package bisq.core.app;
 import bisq.core.arbitration.ArbitratorManager;
 import bisq.core.btc.BtcOptionKeys;
 import bisq.core.btc.RegTestHost;
+import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
-import bisq.core.btc.wallet.WalletsSetup;
 import bisq.core.dao.DaoOptionKeys;
 import bisq.core.dao.DaoSetup;
 import bisq.core.exceptions.BisqException;
