+++ core/src/main/java/bisq/core/app/misc/ExecutableForAppWithP2p.java
@@ -21,9 +21,9 @@ import bisq.core.app.AppOptionKeys;
 import bisq.core.app.BisqEnvironment;
 import bisq.core.app.BisqExecutable;
 import bisq.core.arbitration.ArbitratorManager;
+import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
-import bisq.core.btc.wallet.WalletsSetup;
 import bisq.core.offer.OpenOfferManager;
 
 import bisq.network.p2p.P2PService;
