+++ core/src/main/java/bisq/core/btc/BitcoinModule.java
@@ -18,12 +18,12 @@
 package bisq.core.btc;
 
 import bisq.core.app.AppOptionKeys;
+import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BsqCoinSelector;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.NonBsqCoinSelector;
 import bisq.core.btc.wallet.TradeWalletService;
-import bisq.core.btc.wallet.WalletsSetup;
 import bisq.core.provider.PriceNodeHttpClient;
 import bisq.core.provider.ProvidersRepository;
 import bisq.core.provider.fee.FeeProvider;
