+++ core/src/main/java/bisq/core/btc/setup/ProxySocketFactory.java
@@ -33,7 +33,7 @@
  * PircBotX. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc;
+package bisq.core.btc.setup;
 
 import javax.net.SocketFactory;
