+++ core/src/main/java/bisq/core/btc/wallet/BtcWalletService.java
@@ -24,6 +24,7 @@ import bisq.core.btc.InsufficientFundsException;
 import bisq.core.btc.Restrictions;
 import bisq.core.btc.exceptions.TransactionVerificationException;
 import bisq.core.btc.exceptions.WalletException;
+import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.provider.fee.FeeService;
 import bisq.core.user.Preferences;
