+++ desktop/src/main/java/bisq/desktop/main/funds/withdrawal/WithdrawalView.java
@@ -31,8 +31,8 @@ import bisq.core.btc.AddressEntryException;
 import bisq.core.btc.InsufficientFundsException;
 import bisq.core.btc.Restrictions;
 import bisq.core.btc.listeners.BalanceListener;
+import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BtcWalletService;
-import bisq.core.btc.wallet.WalletsSetup;
 import bisq.core.locale.Res;
 import bisq.core.trade.Tradable;
 import bisq.core.trade.Trade;
